WIP: store data

+++ models/import_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Numeric
+from sqlalchemy import Column, Integer, String, Numeric, BigInteger
 
 from .base import Base
 
@@ -9,5 +9,5 @@ class Import(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     year = Column(Integer, nullable=False)
     country = Column(String, nullable=False)
-    quantity = Column(Integer, nullable=True)
+    quantity = Column(BigInteger, nullable=True)
     value = Column(Numeric(10, 2), nullable=True)
