+++ models/processing_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, BigInteger
 
 from .base import Base
 
@@ -9,4 +9,4 @@ class Processing(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     year = Column(Integer, nullable=False)
     variety = Column(String, nullable=False)
-    quantity = Column(Integer, nullable=True)
+    quantity = Column(BigInteger, nullable=True)
